change stock data type and multiple selected menus

+++ app/src/main/java/com/example/wp/data/api/model/request/RequestMaterialMenuApi.kt
@@ -8,5 +8,5 @@ data class RequestMaterialMenuApi(
     @SerializedName("menu_id")
     val menuId: Int,
     @SerializedName("stock_required")
-    val stockRequired: Int
+    val stockRequired: Double
 )
