+++ app/src/main/java/com/example/wp/domain/material/MaterialMenu.kt
@@ -8,6 +8,6 @@ data class MaterialMenu(
     val id: Int = 0,
     val materialId: Int,
     val menuId: Int,
-    val stockRequired: Int,
+    val stockRequired: Double,
     val material: Material = Material()
 ) : Parcelable
